+++ mcm-app/app/screens/MaterialesScreen.tsx
@@ -130,7 +130,6 @@ export default function MaterialesScreen() {
     ) {
       setIndex(getClosestDateIndex(materialesData));
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [materialesData, route.params?.initialDayIndex]);
   const fechas = materialesData
     ? materialesData.map((d) => ({ fecha: d.fecha }))